+++ R/convert_ket2DM.R
@@ -1,0 +1,4 @@
+
+convert_ket2DM <- function(v){
+	kronecker(v,adjoint(v))
+}