+++ R/many.R
@@ -1,8 +1,9 @@
 #' @export
 many <- function(gate,n,...){	
 	g <- gate
-	for(j in 2:n)
-		g <- tensor(g,gate)	
+	if(n > 1)
+		for(j in 2:n)
+			g <- tensor(g,gate)	
 
 	ket <- list(...)
 	if(length(ket) == 0){