+++ R/opDM.R
@@ -1,0 +1,4 @@
+
+opDM <- function(V,G){
+	G %*% V %*% adjoint(G)
+}